refactor(referentiels): aligne merges statuts/commentaires sur des règles pures
Extrait la projection scoring depuis ScoresService et remplace MergeStatutsService/MergeCommentairesService par mergeStatuts(ctx) et mergeCommentaires(ctx), en préparation de l'orchestration persistance PR17b.
